Add tests for node meta in prepared queries and update docs
