+++ command/agent/agent_test.go
@@ -19,7 +19,6 @@ import (
 	"github.com/hashicorp/consul/logger"
 	"github.com/hashicorp/consul/testutil"
 	"github.com/hashicorp/raft"
-	"strings"
 )
 
 const (
@@ -1849,69 +1848,6 @@ func TestAgent_purgeCheckState(t *testing.T) {
 	file := filepath.Join(agent.config.DataDir, checkStateDir, stringHash("check1"))
 	if _, err := os.Stat(file); !os.IsNotExist(err) {
 		t.Fatalf("should have removed file")
-	}
-}
-
-func TestAgent_metadata(t *testing.T) {
-	// Load a valid set of key/value pairs
-	meta := map[string]string{
-		"key1": "value1",
-		"key2": "value2",
-	}
-	// Should succeed
-	if err := validateMetadata(meta); err != nil {
-		t.Fatalf("err: %s", err)
-	}
-
-	// Should get error
-	meta = map[string]string{
-		"": "value1",
-	}
-	if err := validateMetadata(meta); !strings.Contains(err.Error(), "Couldn't load metadata pair") {
-		t.Fatalf("should have failed")
-	}
-
-	// Should get error
-	meta = make(map[string]string)
-	for i := 0; i < metaMaxKeyPairs+1; i++ {
-		meta[string(i)] = "value"
-	}
-	if err := validateMetadata(meta); !strings.Contains(err.Error(), "cannot contain more than") {
-		t.Fatalf("should have failed")
-	}
-}
-
-func TestAgent_validateMetaPair(t *testing.T) {
-	longKey := strings.Repeat("a", metaKeyMaxLength+1)
-	longValue := strings.Repeat("b", metaValueMaxLength+1)
-	pairs := []struct {
-		Key   string
-		Value string
-		Error string
-	}{
-		// valid pair
-		{"key", "value", ""},
-		// invalid, blank key
-		{"", "value", "cannot be blank"},
-		// allowed special chars in key name
-		{"k_e-y", "value", ""},
-		// disallowed special chars in key name
-		{"(%key&)", "value", "invalid characters"},
-		// key too long
-		{longKey, "value", "Key is too long"},
-		// reserved prefix
-		{metaKeyReservedPrefix + "key", "value", "reserved for internal use"},
-		// value too long
-		{"key", longValue, "Value is too long"},
-	}
-
-	for _, pair := range pairs {
-		err := validateMetaPair(pair.Key, pair.Value)
-		if pair.Error == "" && err != nil {
-			t.Fatalf("should have succeeded: %v, %v", pair, err)
-		} else if pair.Error != "" && !strings.Contains(err.Error(), pair.Error) {
-			t.Fatalf("should have failed: %v, %v", pair, err)
-		}
 	}
 }
