+++ consul/prepared_query/template_test.go
@@ -39,9 +39,9 @@ var (
 				"${match(2)}",
 			},
 			NodeMeta: map[string]string{
-				"${name.full}": "${name.prefix}",
-				"${name.suffix}": "${match(0)}",
-				"${match(1)}": "${match(2)}",
+				"foo": "${name.prefix}",
+				"bar": "${match(0)}",
+				"baz": "${match(1)}",
 			},
 		},
 	}
@@ -227,7 +227,7 @@ func TestTemplate_Render(t *testing.T) {
 				"${match(4)}",
 				"${40 + 2}",
 			},
-			NodeMeta: map[string]string{"${match(1)}": "${match(2)}"},
+			NodeMeta: map[string]string{"foo": "${match(1)}"},
 		},
 	}
 	ct, err := Compile(query)
@@ -258,7 +258,7 @@ func TestTemplate_Render(t *testing.T) {
 					"",
 					"42",
 				},
-				NodeMeta: map[string]string{"hello": "foo"},
+				NodeMeta: map[string]string{"foo": "hello"},
 			},
 		}
 		if !reflect.DeepEqual(actual, expected) {
@@ -289,7 +289,7 @@ func TestTemplate_Render(t *testing.T) {
 					"",
 					"42",
 				},
-				NodeMeta: map[string]string{"": ""},
+				NodeMeta: map[string]string{"foo": ""},
 			},
 		}
 		if !reflect.DeepEqual(actual, expected) {
